Add --processFields to watch and upload to gate field processing.
