+++ packages/cli-lib/lang/en.lyaml
@@ -690,7 +690,9 @@ en:
           options:
             describe: "Options to pass to fields.js"
           saveOutput:
-            decribe: If true, saves all output from fields.js files as 'fields.output.json'.
+            describe: "If true, saves all output from fields.js files as 'fields.output.json'."
+          processFields:
+            describe: "If true, processes any fields.js files contained in module folder or project root."
         previewUrl: "To preview this theme, visit: {{ previewUrl }}"
         success:
           fileUploaded: "Uploaded file from \"{{ src }}\" to \"{{ dest }}\" in the Design Manager of account {{ accountId }}"
