+++ packages/cli-lib/lib/__tests__/uploadFolder.js
@@ -1,8 +1,4 @@
-const {
-  uploadFolder,
-  getFilesByType,
-  resolvePromises,
-} = require('../uploadFolder');
+const { uploadFolder } = require('../uploadFolder');
 const { upload } = require('../../api/fileMapper');
 const { walk, listFilesInDir } = require('../walk');
 const { createIgnoreFilter } = require('../../ignoreRules');
@@ -67,112 +63,6 @@ describe('uploadFolder', () => {
           qs: { buffer: false, environmentId: 1 },
         });
       });
-    });
-  });
-
-  describe('getFilesByType()', () => {
-    it('finds and converts fields.js files to field.json', async () => {
-      const files = [
-        'folder/templates/blog.html',
-        'folder/css/file.css',
-        'folder/js/file.js',
-        'folder/fields.js',
-        'folder/images/image.png',
-        'folder/images/image.jpg',
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.js',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-        'folder/templates/page.html',
-      ];
-
-      listFilesInDir.mockReturnValue(['fields.json']);
-      const [filesByType, compiledJsonFiles] = await resolvePromises(
-        getFilesByType(files, 'folder')
-      );
-      expect(filesByType[1]).toEqual([
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.converted.json',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ]);
-      expect(filesByType[4]).toContain('folder/fields.converted.json');
-      expect(compiledJsonFiles).toEqual(
-        expect.arrayContaining([
-          'folder/fields.converted.json',
-          'folder/sample.module/fields.converted.json',
-        ])
-      );
-    });
-
-    it('does not add field.json files from module folder if a field.js is present in module folder', async () => {
-      const files = [
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.json',
-        'folder/sample.module/fields.js',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ];
-      listFilesInDir.mockImplementation(dir => {
-        return dir === 'folder/sample.module' ? ['fields.js'] : [''];
-      });
-
-      const [filesByType] = await resolvePromises(
-        getFilesByType(files, 'folder')
-      );
-      expect(filesByType[1]).toEqual([
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.converted.json',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ]);
-    });
-
-    it('converts fields.js in root and skips field.json in root', async () => {
-      const files = [
-        'folder/fields.js',
-        'folder/fields.json',
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.json',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ];
-      listFilesInDir.mockImplementation(dir => {
-        return dir === 'folder' ? ['fields.js', 'fields.json'] : [''];
-      });
-
-      const [filesByType, compiledJsonFiles] = await resolvePromises(
-        getFilesByType(files, 'folder')
-      );
-      expect(filesByType[4]).toEqual(['folder/fields.converted.json']);
-      expect(compiledJsonFiles).toEqual(['folder/fields.converted.json']);
-    });
-
-    it('does not add any json files inside of module folder besides fields.json and meta.json', async () => {
-      const files = [
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.json',
-        'folder/sample.module/dont_add.json',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ];
-      listFilesInDir.mockReturnValue(['']);
-      const [filesByType] = await resolvePromises(
-        getFilesByType(files, 'folder')
-      );
-      expect(filesByType[1]).toEqual([
-        'folder/sample.module/module.css',
-        'folder/sample.module/module.js',
-        'folder/sample.module/fields.json',
-        'folder/sample.module/meta.json',
-        'folder/sample.module/module.html',
-      ]);
     });
   });
 });
